backend/requirements.txt
crewai
langchain
duckduckgo-search
python-dotenv
requests
python-docx
reportlab
bcrypt
litellm
huggingface_hub
fastapi
httpx
pydantic-settings
email-validator
uvicorn
